Fix assets paths if they are a URL
If the asset filename is a URL (S3 bucket, Forge Assets library for example),
then it won't match the packPath prefix, so it should use the original
url instead of the removing the url path that comes before it.
It should return the non decoded uri itself in that case.
Also, cache the decoded packPath uri and use slice instead of split
so it's slightly more optimized